change team counter data type

+++ my_log/models.py
@@ -1,4 +1,4 @@
 from django.db import models
 
 class Counter(models.Model):                
-    TeamCount = models.SmallIntegerField(default=0)
+    TeamCount = models.PositiveIntegerField(default=0)
